Improved Model Logic + moar getters & setters
- currentUser now include the role and the school of given user.
- getRoles() is now getRole() because we consider the first Role to be
the topmost: i.e a "teacher" has privileges of a "student" but is first
and foremost a "teacher".
- getSchools() is now getSchool() because we consider (at least
temporarily) that a user only have one school for this email address he
logged in with.
- functions from UserFactory now query the inherited class instead of
the default one to allow getter, setter and inherited functions to take
effect. i.e.: new Parse.Query(RoleFactory) instead of
Parse.Query(Parse.Role).
- binding now works in templates. (no more get() but direct use of
properties)