Parse.Lesson attrs are now pointer (vs relation)
Backend-wise, for the Parse.Lesson Class:
- room, speaker, and course are now pointers instead of relation
because a lesson will have only one of each
- "speakers" in now "speaker" because we consider there is only one
speaker per lesson (at least now)
- attendees becomes a pointer because an attendee is a group
(Parse.Role) of users to attend a lesson. In the case of nested groups,
Parse.Role can already deal with that.

On the Front-end changes are made to reflect this change
(createLessonCtrl).
Also studlife now pulls the lessons of the user group and display the
associated lesson cards.